app: vrbox: door: Fix possible memory consistency problem
When the infrared irq enabled, the irq handler will write true into one
of the infrared detected flag. This flag array will be checked by the
main handle thread, which may cause a memory consistency problem when
access it twice.

Fix this problem by replace the polling method with semaphore to sync
two threads main handle thread and irq handler. To achive this, we have
to change a new API to door open/close function.

All basic functions are tested include the infrared detector. It may not
conver all possible situations, but it still better than previous
version of this code.

Note: The admin_open() and admin_close() are never tested since this
commit. The door open and close pin triggered when its got long trigger
time, it MAY not be triggered if the trigger time is too short.

+++ samples/net/vrbox/src/door.h
@@ -40,24 +40,27 @@ const bool* door_get_status(void);
  *
  * @param layer Door number, [1-4] is expected
  *
- * @return -1 Door has never opened
- *			0 Door has fully opened
- *		   -2 Door open timeout, and leave it half open
+ * @return	 0 Door has fully opened
+ *			-1 Door open timeout
+ *			-2 Door never opened
+ *			-3 Door half opened
+ *			-4 Door is at opened AND closed state, impossible
  * */
-int8_t door_open(uint8_t layer);
+int door_open(uint8_t layer);
 
 /**
  * @brief Public wrapper function of door close
  *
  * @param layer Door number, [1-4] is expected
  *
- * @return -3 Door has never close after timeout
- *		   -2 Door has been blocked by something which detected by on
- *			  door infrared detector
- *		   -1 Door close timeout, and leave it half close
- *			0 Door has fully closed
+ * @return	 1 On door infrared detected
+ *			 0 Door has fully closed
+ *			-1 Door close timeout
+ *			-2 Door never closed
+ *			-3 Door half closed
+ *			-4 Door is at opened AND closed state, impossible
  * */
-int8_t door_close(uint8_t layer);
+int door_close(uint8_t layer);
 
 /**
  * @brief Open all four doors at a time
